fixing imports and renaming PostDetails and PostDetailsComment
